aws-eks-platform-terraform/app/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
boto3==1.34.34
python-dotenv==1.0.0
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
pydantic==2.5.3
pydantic-settings==2.1.0
